+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## Version [4.0.0](https://github.com/cedx/sql.net/compare/v3.2.0...v4.0.0)
 - Breaking change: raised the required target framework.
 - Breaking change: removed the `Invoke-Reader` cmdlet.
+- Breaking change: renamed the `QueryOptions.Buffered` property to `Stream`.
 - Ported the [C#](https://learn.microsoft.com/en-us/dotnet/csharp) cmdlets to [PowerShell](https://learn.microsoft.com/en-us/powershell).
 
 ## Version [3.2.0](https://github.com/cedx/sql.net/compare/v3.1.0...v3.2.0)